feat: Create `GoogleCallbackClient` component to handle Google OAuth token processing and update `GoogleCallbackPage` to utilize it for improved code organization.
